DAOS-10936 ddb: First DDB JSON output
Added json output for the superblock dump command. In order to
accomplish this using golang for the json output, had to move the
printing for the command into the go code. The command now calls
directly into the ddb_vos layer instead of the ddb_command layer of
the C library.

Added a commands context structure to wrap the C structure to make it
easier to pass around command line options.

Will apply similar changes to the other commands when this change is
reviewed  and approved.

Test-tag: ddb_cmd

Required-githooks: true

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>